feat: 유효성 검사 로직 추가(#26)
